feat: extract component
